requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
streamlit
